requirements.txt
alpaca-trade-api 
psycopg2-binary
